requirements.txt
annotated-types==0.7.0
anyio==4.10.0
blinker==1.9.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.3.0
colorama==0.4.6
fastapi==0.117.1
Flask==3.1.2
h11==0.16.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
pydantic==2.11.9
pydantic_core==2.33.2
python-dotenv==1.1.1
redis==6.4.0
requests==2.32.5
sniffio==1.3.1
spotipy==2.25.1
starlette==0.48.0
typing-inspection==0.4.1
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.36.0
Werkzeug==3.1.3
